+++ tests/files/src/0800_pass_by_ref.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace N\S800;
+
+function test_assign(&$x) {
+    [$x] = [new \stdClass()];
+}
+test_assign($globalVar);
+echo count($globalVar);