+++ src/Phan/Config.php
@@ -854,6 +854,13 @@ class Config
 
         'dump_matching_functions' => false,
 
+        // This is the path to a file containing a list of pre-existing issues to ignore, on a per-file basis.
+        // It's recommended to set this with --load-baseline=path/to/baseline.php
+        'baseline_path' => null,
+
+        // For internal use only.
+        '__save_baseline_path' => null,
+
         // A list of plugin files to execute.
         //
         // Plugins which are bundled with Phan can be added here by providing their name (e.g. `'AlwaysReturnPlugin'`)
@@ -1287,6 +1294,7 @@ class Config
             'array_casts_as_null' => $is_bool,
             'autoload_internal_extension_signatures' => $is_associative_string_array,
             'backward_compatibility_checks' => $is_bool,
+            'baseline_path' => $is_string_or_null,
             'cache_polyfill_asts' => $is_bool,
             'check_docblock_signature_param_type_match' => $is_bool,
             'check_docblock_signature_return_type_match' => $is_bool,