+++ NEWS.md
@@ -13,6 +13,7 @@ New features(Analysis):
   This helps analyze `PassByReferenceVariable`s.
 + Don't emit PhanParamSpecial2 for min/max/implode/join with a single vararg. (#3396)
 + Properly emit PhanPossiblyInfiniteRecursionSameParams for functions with varargs.
++ Emit `PhanNoopNew` or `PhanNoopNewNoSideEffects` when an object is created with `new expr(...)` but the result is not used (#3410)
 
 Oct 20 2019, Phan 2.3.1
 -----------------------