+++ NEWS.md
@@ -14,6 +14,7 @@ New features(Analysis):
 + Don't emit PhanParamSpecial2 for min/max/implode/join with a single vararg. (#3396)
 + Properly emit PhanPossiblyInfiniteRecursionSameParams for functions with varargs.
 + Emit `PhanNoopNew` or `PhanNoopNewNoSideEffects` when an object is created with `new expr(...)` but the result is not used (#3410)
++ Emit `PhanPluginUseReturnValueInternalKnown` for about unused results of function calls on the right hand side of control flow operators (`??`/`?:`/`&&`/`||`) (#3408)
 
 Oct 20 2019, Phan 2.3.1
 -----------------------